OBS_TBot/requirements.txt
flask==2.1.0
aiogram==2.24
requests
python-dotenv
# pytelegrambotapi==4.0.0

# Для гугл таблиц
# gspread 
# oauth2client
# google-api-python-client
gspread
google-auth

# для okato-oktmo
# pandas
# xlsxwriter